refactoring withdraw to float32

+++ internal/entity/BalanceRequest.go
@@ -1,6 +1,6 @@
 package entity
 
 type BalanceRequest struct {
-	Current   float64 `json:"current,omitempty"`
-	Withdrawn float64 `json:"withdrawn,omitempty"`
+	Current   float32 `json:"current,omitempty"`
+	Withdrawn float32 `json:"withdrawn,omitempty"`
 }
