+++ internal/entity/myerrors/limitExceededError.go
@@ -5,8 +5,8 @@ import (
 )
 
 type LimitExceededError struct {
-	Sum          float64
-	TotalAccrual float64
+	Sum          float32
+	TotalAccrual float32
 	UserID       string
 }
 
@@ -14,7 +14,7 @@ func (le *LimitExceededError) Error() string {
 	return fmt.Sprintf("sum %f exceededs limit, for user %s accessible limit is %f", le.Sum, le.UserID, le.TotalAccrual)
 }
 
-func NewLimitExceededError(sum float64, totalAccrual float64, userID string) error {
+func NewLimitExceededError(sum float32, totalAccrual float32, userID string) error {
 	return &LimitExceededError{
 		Sum:          sum,
 		TotalAccrual: totalAccrual,
